Moved all components to a new file
