+++ src/app/layout.tsx
@@ -1,13 +1,12 @@
-import Link from "next/link";
-import Image from "next/image";
 import { Outfit } from 'next/font/google'
-import { ReactNode } from "react";
- 
+import { SideBar } from "../components/SideBar";
+
+import "./style.css";
+
 const outfit = Outfit({
   subsets: ['latin'],
 })
- 
-import "./style.css";
+
 
 export default function RootLayout({
   children,
@@ -16,44 +15,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" className={outfit.className}>
-      <body>
+      <body className="flex w-screen h-screen">
         <SideBar />
-        <Body>
-          {children}
-        </Body>
+        {children}
       </body>
     </html>
   );
-}
-
-
-function Body({ children }: { children: ReactNode }) {
-  return <Section id="body" className="w-full p-6">{children}</Section>
-}
-function Section({ id, className, children }: { id: string, className?: string, children: React.ReactNode }) {
-  return <div id={id} className={`bg-neutral-900 m-4 rounded-3xl p-2 ${className}`}>
-    {children}
-  </div>
-}
-
-function SideBar() {
-  return <Section id="sidebar" className="flex flex-col items-center">
-    {
-      [
-        { id: "home", data: { icon: "/svgs/home.svg", path: "/", label: "Home" } },
-        { id: 'test', data: { icon: "/svgs/test.svg", path: "/test", label: "Test" } },
-        { id: 'ex', data: { icon: "/svgs/exercices.svg", path: "/exercices", label: "Exercices" } }
-      ]
-        .map(btn => <SideBarIcon {...btn.data} key={btn.id} />)
-    }
-  </Section>
-}
-
-function SideBarIcon({ icon, path, label }: { icon: string, path: string, label: string }) {
-  return <Link href={path} className="aspect-square rounded-xl 
-  hover:bg-neutral-50/5 hover:scale-105
-  p-2 m-1">
-    <Image src={icon} alt={label} width={40} height={40} />
-  </Link>
-
 }
